New Folder Structure for Homepage
Trying to fix the bug
