--- components/sections/Homepage/BlazeAnnouncement/BlazeAnnouncement.module.scss
@@ -1,31 +1,0 @@
-.sectionContainer {
-  border: 16px solid #e6c619;
-}
-.container {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.text {
-  max-width: 504px;
-}
-
-@media (max-width: 768px) {
-  .container {
-    flex-direction: column;
-  }
-
-  .text {
-    max-width: 400px;
-    text-align: center;
-    margin: 3.5rem;
-  }
-}
-
-@media (max-width: 900px) {
-  .imgBlaze {
-    width: 238px;
-    height: 47px;
-  }
-}
