+++ components/sections/fresh-home/blaze/blaze.module.scss
@@ -1,0 +1,64 @@
+.sectionContainer {
+    border: 16px solid #e6c619;
+}
+
+.container {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+
+.text {
+    max-width: 504px;
+    line-height: 125%;
+}
+
+.text a {
+    border-bottom: 3px solid;
+    padding-bottom: 2px;
+}
+
+//RESPONSIVE STYLING SECTION
+@media (max-width: 1000px) {
+    .text {
+        font-size: 3.6rem;
+        line-height: 125%;
+        text-align: left;
+        max-width: 350px;
+    }
+}
+
+@media screen and (min-width: 876px) and (max-width: 999px) {
+    .text {
+        font-size: 3rem;
+        line-height: 150%;
+        padding: 2.6rem 0;
+        max-width: 400px;
+    }
+}
+
+@media screen and (min-width: 769px) and (max-width: 875px) {
+    .text {
+        font-size: 2.4rem;
+        padding: 2.6rem 0;
+        max-width: 400px;
+    }
+}
+
+@media (max-width: 768px) {
+    .container {
+        flex-direction: column;
+    }
+
+    .text {
+        font-size: 3.2rem;
+        line-height: 125%;
+        text-align: center;
+        max-width: 400px;
+    }
+
+    .text a {
+        border-bottom: 2px solid;
+        padding-bottom: 1px;
+    }
+}
