+++ components/sections/fresh-home/programmes/programmes.module.scss
@@ -1,0 +1,93 @@
+@import "../../../../styles/variables.scss";
+
+.sectionContainer {
+    display: flex;
+    flex-direction: column;
+    gap: 4rem;
+}
+
+.headingContainer {
+    display: flex;
+    align-items: center;
+    justify-content: space-around;
+    gap: 1rem;
+}
+
+.imageContainer {
+    padding-right: 2.2rem;
+    padding-bottom: 2.4rem;
+}
+
+.imageContainer>span {
+    overflow: visible;
+    box-shadow: 22px 24px $PRIMARY_HONEY;
+}
+
+.headingText {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: 50%;
+    padding-left: 1rem;
+    max-width: 350px;
+    text-align: center;
+}
+
+.sectionDivider {
+    width: 100%;
+    height: 1px;
+    background-color: $GRAY_3;
+}
+
+.moreInfoContainer {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    gap: 21.6rem;
+}
+
+.infoWrapper {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    width: 305px;
+}
+
+@media (max-width: 992px) {
+    .moreInfoContainer {
+        gap: 12rem;
+    }
+}
+
+@media (max-width: 900px) {
+    .moreInfoContainer {
+        gap: 6rem;
+    }
+}
+
+@media (max-width: 730px) {
+    .sectionContainer {
+        gap: 2rem;
+    }
+
+    .headingContainer {
+        flex-direction: column;
+        gap: 0;
+    }
+
+    .headingText {
+        width: 100%;
+        padding: 2rem 0 0;
+    }
+
+    .moreInfoContainer {
+        flex-direction: column;
+        padding-top: 1rem;
+        gap: 3rem;
+    }
+
+    .sectionDivider {
+        display: none;
+    }
+}
