--- components/sections/homepage/BlazeAnnouncement/BlazeAnnouncement.module.scss
@@ -1,59 +1,0 @@
-.sectionContainer {
-  border: 16px solid #e6c619;
-}
-
-.container {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.text {
-  max-width: 504px;
-}
-
-//WAITING FOR UX DESIGN TEAM IF THEY WANT AN UNDERLINE. AT THE MEANTIME, HERE'S THE UNDERLINE
-.text a {
-  border-bottom: 3px solid;
-  padding-bottom: 2px;
-}
-
-@media (max-width: 1000px) {
-  .text {
-    font-size: 3.6rem;
-    line-height: 100%;
-    text-align: left;
-    max-width: 350px;
-  }
-}
-
-//MAKE THE FRIGGIN BLAZE BANNER RESPONSIVE AAAAAAAAAAAAAAAAAAAA
-/*@media (max-width: 925px) {
-  .text {
-    font-size: 3.2rem;
-    padding: 2.6rem 0;
-    max-width: 400px;
-  }
-}
-*/
-
-/*@media (min-width: 769px) {
-  .text {
-    font-size: 2.6rem;
-    padding: 2.6rem 0;
-    max-width: 400px;
-   }
-}
-*/
-
-/* @media (max-width: 768px) {
-  .container {
-    flex-direction: column;
-  }
-  .text {
-    font-size: 3.2rem;
-    text-align: center;
-    max-width: 400px;
- }
-}
-*/
