--- components/sections/homepage/Career/Career.module.scss
@@ -1,27 +1,0 @@
-.text {
-  max-width: 630px;
-  margin: 1rem auto;
-  text-align: center;
-}
-.italics {
-  font-style: italic;
-}
-
-.cards {
-  margin-top: 6rem;
-  display: flex;
-  align-items: center;
-}
-
-@media (max-width: 992px) {
-  .cards > div > div {
-    max-width: 215px;
-  }
-}
-
-@media (max-width: 768px) {
-  .cards {
-    flex-direction: column;
-    margin-top: 2rem;
-  }
-}
